user-profile: handle authentication to resources

+++ user-profile/src/main/java/com/oxygensened/userprofile/context/education/EducationService.java
@@ -1,40 +1,52 @@
 package com.oxygensened.userprofile.context.education;
 
+import com.oxygensend.commons_jdk.exception.AccessDeniedException;
+import com.oxygensend.commons_jdk.request_context.RequestContext;
 import com.oxygensened.userprofile.context.education.dto.request.CreateEducationRequest;
-import com.oxygensened.userprofile.context.education.dto.view.EducationView;
 import com.oxygensened.userprofile.context.education.dto.request.UpdateEducationRequest;
+import com.oxygensened.userprofile.context.education.dto.view.EducationView;
+import com.oxygensened.userprofile.context.utils.JsonNullableWrapper;
 import com.oxygensened.userprofile.domain.entity.Education;
-import com.oxygensened.userprofile.domain.repository.EducationRepository;
 import com.oxygensened.userprofile.domain.entity.University;
-import com.oxygensened.userprofile.domain.repository.UniversityRepository;
-import com.oxygensened.userprofile.domain.repository.UserRepository;
 import com.oxygensened.userprofile.domain.exception.EducationNotFoundException;
 import com.oxygensened.userprofile.domain.exception.NoSuchUniversityException;
 import com.oxygensened.userprofile.domain.exception.UserNotFoundException;
-import com.oxygensened.userprofile.context.utils.JsonNullableWrapper;
+import com.oxygensened.userprofile.domain.repository.EducationRepository;
+import com.oxygensened.userprofile.domain.repository.UniversityRepository;
+import com.oxygensened.userprofile.domain.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.function.Consumer;
 import org.openapitools.jackson.nullable.JsonNullable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import static com.oxygensened.userprofile.context.utils.Patch.updateIfPresent;
 
 @Service
 public class EducationService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(EducationService.class);
     private final EducationRepository educationRepository;
     private final UniversityRepository universityRepository;
     private final UserRepository userRepository;
+    private final RequestContext requestContext;
 
-    public EducationService(EducationRepository educationRepository, UniversityRepository universityRepository, UserRepository userRepository) {
+    public EducationService(EducationRepository educationRepository, UniversityRepository universityRepository, UserRepository userRepository, RequestContext requestContext) {
         this.educationRepository = educationRepository;
         this.universityRepository = universityRepository;
         this.userRepository = userRepository;
+        this.requestContext = requestContext;
     }
 
     public EducationView createEducation(Long userId, CreateEducationRequest request) {
         var user = userRepository.findById(userId)
                                  .orElseThrow(() -> UserNotFoundException.withId(userId));
+
+        if (!requestContext.isUserAuthenticated(userId)) {
+            LOGGER.info("User {} is not allow to create education for different entities", requestContext.userIdAsString());
+            throw new AccessDeniedException();
+        }
 
         var university = universityRepository.findById(request.universityId())
                                              .orElseThrow(() -> NoSuchUniversityException.withId(request.universityId()));
@@ -58,6 +70,11 @@ public class EducationService {
         var education = educationRepository.findByIdAndUserId(educationId, userId)
                                            .orElseThrow(() -> EducationNotFoundException.withId(educationId));
 
+        if (!requestContext.isUserAuthenticated(userId)) {
+            LOGGER.info("User {} is not allowed to update education {}", requestContext.userIdAsString(), educationId);
+            throw new AccessDeniedException();
+        }
+
         updateUniversity(request.universityId(), education::setUniversity);
         updateIfPresent(request.fieldOfStudy(), education::setFieldOfStudy);
         updateIfPresent(request.title(), education::setTitle);
@@ -74,6 +91,10 @@ public class EducationService {
     public void deleteEducation(Long userId, Long educationId) {
         var education = educationRepository.findByIdAndUserId(educationId, userId)
                                            .orElseThrow(() -> EducationNotFoundException.withId(educationId));
+        if (!requestContext.isUserAuthenticated(userId)) {
+            LOGGER.info("User {} is not allowed to delete education {}", requestContext.userIdAsString(), educationId);
+            throw new AccessDeniedException();
+        }
 
         educationRepository.delete(education);
     }
